requirements.txt
attrs==20.3.0
azure-functions==1.5.0
click==7.1.2
coverage==5.5
Flask==1.1.2
importlib-metadata==3.10.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
toml==0.10.2
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.1
